LeaveFlow-main/backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - use compatible versions for cloud deployment
sqlalchemy==2.0.25
asyncpg==0.28.0
aiosqlite==0.19.0
alembic==1.13.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Explicit version compatible with passlib
pydantic==2.10.3  # Updated to avoid Rust compilation issues
pydantic-settings==2.1.0
email-validator==2.1.0  # Required by Pydantic for EmailStr validation

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# AI - OpenRouter for free LLM access (multiple free models)
openai==1.12.0  # OpenRouter uses OpenAI-compatible API

# Date parsing
python-dateutil==2.8.2

# Scheduling
apscheduler==3.10.4

# Storage (optional) - Commented out for serverless
# python-magic==0.4.27

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
